Override xdg-open to suppress benign error
PIVOT uses the launch.browser option for Shiny, which is incompatible
with Docker, leading to an xdg-open error. We override this with a fake
xdg-open command that echoes a useful message to the user. However, it
seems the message is not passed to the terminal, though the error is
indeed suppressed. This is merely a quick workaround until a good way of
turning off the switch and passing a start message is implemented.